examples/agents/python/adk-quickstart/pyproject.toml
[project]
name = "adk-quickstart"
version = "0.1.0"
description = "Minimal Google ADK agent deployed on GravixLayer."
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.0.0",
    "a2a-sdk[http-server]>=1.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["time_agent*"]
